refactor: move a320 press code out of systems
