+++ fbw-a380x/src/wasm/systems/a380_systems/src/lib.rs
@@ -9,7 +9,7 @@ mod pneumatic;
 mod power_consumption;
 
 use self::{
-    air_conditioning::A380AirConditioning,
+    air_conditioning::{A380AirConditioning, A380PressurizationOverheadPanel},
     fuel::A380Fuel,
     pneumatic::{A380Pneumatic, A380PneumaticOverheadPanel},
 };
@@ -35,7 +35,6 @@ use systems::{
     navigation::adirs::{
         AirDataInertialReferenceSystem, AirDataInertialReferenceSystemOverheadPanel,
     },
-    pressurization::{Pressurization, PressurizationOverheadPanel},
     shared::ElectricalBusType,
     simulation::{Aircraft, SimulationElement, SimulationElementVisitor, UpdateContext},
 };
@@ -48,6 +47,7 @@ pub struct A380 {
     apu_fire_overhead: AuxiliaryPowerUnitFireOverheadPanel,
     apu_overhead: AuxiliaryPowerUnitOverheadPanel,
     pneumatic_overhead: A380PneumaticOverheadPanel,
+    pressurization_overhead: A380PressurizationOverheadPanel,
     electrical_overhead: A380ElectricalOverheadPanel,
     emergency_electrical_overhead: A380EmergencyElectricalOverheadPanel,
     fuel: A380Fuel,
@@ -64,8 +64,6 @@ pub struct A380 {
     hydraulic_overhead: A380HydraulicOverheadPanel,
     autobrake_panel: AutobrakePanel,
     landing_gear: LandingGear,
-    pressurization: Pressurization,
-    pressurization_overhead: PressurizationOverheadPanel,
     pneumatic: A380Pneumatic,
     radio_altimeters: A380RadioAltimeters,
     engines_flex_physics: EnginesFlexiblePhysics<4>,
@@ -86,6 +84,7 @@ impl A380 {
             apu_fire_overhead: AuxiliaryPowerUnitFireOverheadPanel::new(context),
             apu_overhead: AuxiliaryPowerUnitOverheadPanel::new(context),
             pneumatic_overhead: A380PneumaticOverheadPanel::new(context),
+            pressurization_overhead: A380PressurizationOverheadPanel::new(context),
             electrical_overhead: A380ElectricalOverheadPanel::new(context),
             emergency_electrical_overhead: A380EmergencyElectricalOverheadPanel::new(context),
             fuel: A380Fuel::new(context),
@@ -106,8 +105,6 @@ impl A380 {
             hydraulic_overhead: A380HydraulicOverheadPanel::new(context),
             autobrake_panel: AutobrakePanel::new(context),
             landing_gear: LandingGear::new(context),
-            pressurization: Pressurization::new(context),
-            pressurization_overhead: PressurizationOverheadPanel::new(context),
             pneumatic: A380Pneumatic::new(context),
             radio_altimeters: A380RadioAltimeters::new(context),
             engines_flex_physics: EnginesFlexiblePhysics::new(context),
@@ -167,13 +164,6 @@ impl Aircraft for A380 {
 
         self.radio_altimeters.update(context);
 
-        self.pressurization.update(
-            context,
-            &self.pressurization_overhead,
-            [&self.engine_1, &self.engine_2],
-            [self.lgcius.lgciu1(), self.lgcius.lgciu2()],
-        );
-
         self.hydraulic.update(
             context,
             [
@@ -210,6 +200,8 @@ impl Aircraft for A380 {
             &self.apu,
             &self.air_conditioning,
         );
+        self.air_conditioning
+            .mix_packs_air_update(self.pneumatic.packs());
         self.air_conditioning.update(
             context,
             &self.adirs,
@@ -217,7 +209,6 @@ impl Aircraft for A380 {
             &self.engine_fire_overhead,
             &self.pneumatic,
             &self.pneumatic_overhead,
-            &self.pressurization,
             &self.pressurization_overhead,
             [self.lgcius.lgciu1(), self.lgcius.lgciu2()],
         );
@@ -237,6 +228,7 @@ impl SimulationElement for A380 {
         self.emergency_electrical_overhead.accept(visitor);
         self.fuel.accept(visitor);
         self.pneumatic_overhead.accept(visitor);
+        self.pressurization_overhead.accept(visitor);
         self.engine_1.accept(visitor);
         self.engine_2.accept(visitor);
         self.engine_3.accept(visitor);
@@ -251,8 +243,6 @@ impl SimulationElement for A380 {
         self.hydraulic.accept(visitor);
         self.hydraulic_overhead.accept(visitor);
         self.landing_gear.accept(visitor);
-        self.pressurization.accept(visitor);
-        self.pressurization_overhead.accept(visitor);
         self.pneumatic.accept(visitor);
         self.engines_flex_physics.accept(visitor);
